+++ resources/views/relatorios/movimentacoes/partials/termo-responsabilidade-funcionario.blade.php
@@ -122,7 +122,7 @@
         <tr>
             <th>Item</th>
             <th>ID Equip.</th>
-            <th>Descrição</th>
+            <th class="texto-esquerda">Equipamento</th>
             <th>Nº de Série</th>
             <th>Patrimônio</th>
             <th>Valor de Compra (R$)</th>
@@ -137,9 +137,7 @@
             <tr>
                 <td>{{ $numeroItem }}</td>
                 <td>{{ $equipamento->id }}</td>
-                <td class="texto-esquerda">
-                    {{ $equipamento->nome_exibicao ?: 'Descrição não informada' }}
-                </td>
+                <td class="texto-esquerda">@include('relatorios.partials.equipamento-celula')</td>
                 <td>{{ $equipamento->numero_serie ?? ($equipamento->serial ?? '') }}</td>
                 <td>{{ $equipamento->patrimonio ?? ($equipamento->numero_patrimonio ?? '') }}</td>
                 <td>